Some work on the MAPM types (decimal &amp; integer) and the NumConversions class 1 Cleanup 2 Some perf improvements 3 Some new numeric conversions 4 Some bug fixes (fixed 4 XQTS tests: K2-SeqExprCast-83, K2-SeqExprCast-84,   K2-SeqExprCast-85 and K2-SeqExprCast-428)

+++ src/zorbatypes/duration.h
@@ -184,7 +184,7 @@ protected:
   FACET_TYPE facet;
 
   bool is_negative;
-  long data[7];
+  int32_t data[7];
 public:
   SERIALIZABLE_CLASS(Duration)
   SERIALIZABLE_CLASS_CONSTRUCTOR(Duration)
